requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-community
langchain-text-splitters
langchain-huggingface
pdfplumber
faiss-cpu
sentence-transformers
pydub
openai-whisper
gTTS
sqlalchemy
bcrypt
python-jose[cryptography]
python-multipart
pymysql
